refactor/order: Move processOrder logic to model
